Refactor into class an function modules.
