+++ datashuttle/__init__.py
@@ -1,6 +1,8 @@
 from importlib.metadata import PackageNotFoundError, version
 
-from datashuttle.datashuttle import DataShuttle
+from datashuttle.datashuttle_class import DataShuttle
+from datashuttle.datashuttle_functions import quick_validate_project
+
 
 try:
     __version__ = version("datashuttle")
