+++ docs/source/pages/api_index.rst
@@ -3,6 +3,11 @@
 API Reference
 -----------------
 
-.. currentmodule:: datashuttle.datashuttle
+.. currentmodule:: datashuttle.datashuttle_class
 
 .. autoclass:: DataShuttle
+
+.. automodule:: datashuttle.datashuttle_functions
+   :members:
+   :undoc-members:
+   :show-inheritance:
