+++ tests/tests_integration/base.py
@@ -4,7 +4,7 @@ import warnings
 import pytest
 import test_utils
 
-from datashuttle.datashuttle import DataShuttle
+from datashuttle import DataShuttle
 
 TEST_PROJECT_NAME = "test_project"
